Added more tests and fixed minor bugs;
Added tests for SaleInfo and fixed minor bugs detected by sonarcloud
regarding thread reinterruption and long casting.

+++ crawler/src/main/java/com/javaacademy/crawler/ScrapperScheduledRunner.java
@@ -19,6 +19,6 @@ public class ScrapperScheduledRunner {
     };
 
     public ScrapperScheduledRunner() {
-        timer.schedule (timerTask, 0l, 1000*60*60);
+        timer.schedule (timerTask, 0l, (long)1000*60*60);
     }
 }
